refacto(commons): move `load_from_file` in commons
This utility function was used in several places in the code other than in the
`descriptors`, it made more sense to move it there.
