+++ zenoh-flow-commons/Cargo.toml
@@ -28,6 +28,8 @@ anyhow = { workspace = true }
 bytesize = "1.2.0"
 humantime = "2.1"
 uuid = { workspace = true }
+ramhorns = "0.14"
 serde = { workspace = true }
 serde_json = { workspace = true }
+serde_yaml = { workspace = true }
 zenoh-keyexpr = { workspace = true }
