+++ zenoh-flow-descriptors/Cargo.toml
@@ -25,7 +25,6 @@ version = { workspace = true }
 
 [dependencies]
 anyhow = { workspace = true }
-ramhorns = "0.14"
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_yaml = { workspace = true }
